requirements.txt
# Core dependencies
requests>=2.32.4
flask>=3.1.1
python-dotenv>=1.1.1
aiohttp>=3.9.0
tiktoken>=0.5.0

# Data validation and serialization
pydantic>=2.5.0
marshmallow>=3.20.0

# Logging and monitoring
structlog>=23.2.0
python-json-logger>=2.0.0
prometheus-client>=0.19.0
flask-healthz>=1.0.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
testcontainers>=3.7.0
factory-boy>=3.3.0
responses>=0.24.0

# Code quality and formatting
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
bandit>=1.7.0
safety>=2.3.0
pre-commit>=3.6.0

# Security and authentication
cryptography>=41.0.0
PyJWT>=2.8.0
bcrypt>=4.1.0

# Performance and caching
redis>=5.0.0
gunicorn>=21.2.0
gevent>=23.9.0

# Development tools
ipython>=8.18.0
debugpy>=1.8.0